[Added] changed structure to extra crate filehash.rs
