+++ tests/lib.rs
@@ -18,7 +18,9 @@ fn test_read_from_directory() {
     let input_dir = "tests/example_md/real_md";
     assert!(wiki.read_from_directory(input_dir).is_ok());
     let sha_file = Path::new("html").join(".files.sha");
-    assert!(fs::remove_file(&sha_file).is_ok());
+    if sha_file.exists() {
+        assert!(fs::remove_file(&sha_file).is_ok());
+    }
     assert!(wiki.read_content_from_current_paths(input_dir, "html").is_ok());
     assert!(sha_file.exists());
     assert!(Path::new("html").exists());
@@ -36,7 +38,9 @@ fn test_read_from_directory() {
     println!("The following paths were found:");
     wiki.list_current_input_paths();
     let index_file = Path::new("html").join("index.html");
-    assert!(fs::remove_file(&index_file).is_ok());
+    if index_file.exists() {
+        assert!(fs::remove_file(&index_file).is_ok());
+    }
     assert!(wiki.create_index_tree("html").is_ok());
     assert!(index_file.exists());
 }
@@ -44,10 +48,12 @@ fn test_read_from_directory() {
 #[test]
 fn test_sha_file_existing() {
     let mut wiki = Wiki::new();
-    let sha_file = Path::new("html").join(".files.sha");
-    assert!(sha_file.exists());
     let input_dir = "tests/example_md/real_md";
-    assert!(wiki.read_content_from_current_paths(input_dir, "html").is_ok());
+    assert!(wiki.read_from_directory(input_dir).is_ok());
+    assert!(wiki.read_content_from_current_paths(input_dir, "html2").is_ok());
+    let sha_file = Path::new("html2").join(".files.sha");
+    assert!(sha_file.exists());
+    assert!(wiki.read_content_from_current_paths(input_dir, "html2").is_ok());
 }
 
 #[test]
